services/nemotron_asr/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
numpy==1.26.4
nemo_toolkit[asr]
soundfile
